# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: YouTube Analytics Dashboard API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: channels
    description: Channel analytics
  - name: overview
    description: Aggregated dashboard overview
  - name: platforms
    description: Supported social platforms
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /platforms:
    get:
      operationId: getPlatforms
      tags: [platforms]
      summary: List all supported platforms
      description: Returns metadata for all platforms that can be tracked
      responses:
        "200":
          description: Platform list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Platform"

  /channels:
    get:
      operationId: getChannels
      tags: [channels]
      summary: List all tracked channels
      description: Returns all channels with summary stats
      responses:
        "200":
          description: Channel list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChannelSummary"
    post:
      operationId: createChannel
      tags: [channels]
      summary: Add a new channel to track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChannelRequest"
      responses:
        "201":
          description: Channel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelSummary"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Channel already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /channels/{channelId}:
    get:
      operationId: getChannel
      tags: [channels]
      summary: Get a single channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Channel summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelSummary"
        "404":
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateChannel
      tags: [channels]
      summary: Update a channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChannelRequest"
      responses:
        "200":
          description: Updated channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelSummary"
        "404":
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteChannel
      tags: [channels]
      summary: Remove a channel from tracking
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Channel deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteChannelResponse"
        "404":
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /channels/{channelId}/metrics:
    get:
      operationId: getChannelMetrics
      tags: [channels]
      summary: Get daily metrics for a channel
      description: Returns daily views, subscribers, watch time, and engagement
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
        - name: days
          in: query
          required: false
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Daily metrics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyMetric"
        "404":
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /channels/{channelId}/videos:
    get:
      operationId: getChannelVideos
      tags: [channels]
      summary: Get top videos for a channel
      description: Returns top videos sorted by views
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Video list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VideoRow"
        "404":
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /overview:
    get:
      operationId: getOverview
      tags: [overview]
      summary: Get aggregated overview stats
      description: Returns KPI totals across all channels for the last 30 days
      responses:
        "200":
          description: Overview stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverviewStats"

  /overview/trends:
    get:
      operationId: getOverviewTrends
      tags: [overview]
      summary: Get cross-channel daily trends
      description: Returns aggregated daily metrics across all channels
      parameters:
        - name: days
          in: query
          required: false
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Daily trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendPoint"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Platform:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
        supportsAnalytics:
          type: boolean
        supportsRevenue:
          type: boolean
      required:
        - id
        - name
        - color
        - supportsAnalytics
        - supportsRevenue

    ChannelSummary:
      type: object
      description: |
        Public stats (subscribers, totalViews, totalVideos) come from the
        YouTube Data API and are real for every imported channel. Per-channel
        time-windowed metrics (subscriberGrowth30d, viewsGrowth30d) require
        per-brand OAuth and are null for channels we can't authenticate. The
        engagementRate is computed from real recent video stats via the Data
        API and is non-null whenever recent videos exist.
      properties:
        id:
          type: string
        name:
          type: string
        handle:
          type: string
        platform:
          type: string
        url:
          type: string
        avatarColor:
          type: string
        subscribers:
          type: integer
        totalViews:
          type: integer
        totalVideos:
          type: integer
        totalWatchTimeHours:
          type: number
          nullable: true
        avgViewsPerVideo:
          type: integer
        hasAnalyticsAccess:
          type: boolean
          description: True when a per-brand OAuth token is available for this channel and the Analytics API can be queried for time-windowed metrics.
        subscriberGrowth30d:
          type: integer
          nullable: true
        viewsGrowth30d:
          type: number
          nullable: true
        engagementRate:
          type: number
          nullable: true
      required:
        - id
        - name
        - handle
        - platform
        - avatarColor
        - subscribers
        - totalViews
        - totalVideos
        - avgViewsPerVideo
        - hasAnalyticsAccess

    CreateChannelRequest:
      type: object
      properties:
        name:
          type: string
        handle:
          type: string
        platform:
          type: string
        avatarColor:
          type: string
        url:
          type: string
        subscribers:
          type: integer
        totalViews:
          type: integer
        totalVideos:
          type: integer
      required:
        - name
        - handle
        - platform
        - avatarColor

    UpdateChannelRequest:
      type: object
      properties:
        name:
          type: string
        handle:
          type: string
        platform:
          type: string
        avatarColor:
          type: string
        url:
          type: string
        subscribers:
          type: integer
        totalViews:
          type: integer
        totalVideos:
          type: integer

    DeleteChannelResponse:
      type: object
      properties:
        success:
          type: boolean
        id:
          type: string
      required:
        - success
        - id

    DailyMetric:
      type: object
      properties:
        date:
          type: string
        views:
          type: integer
        subscribers:
          type: integer
        watchTimeHours:
          type: number
        likes:
          type: integer
        comments:
          type: integer
        shares:
          type: integer
        engagementRate:
          type: number
        estimatedRevenue:
          type: number
      required:
        - date
        - views
        - subscribers
        - watchTimeHours
        - likes
        - comments
        - shares
        - engagementRate
        - estimatedRevenue

    VideoRow:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        publishedAt:
          type: string
        views:
          type: integer
        likes:
          type: integer
        comments:
          type: integer
        watchTimeHours:
          type: number
        engagementRate:
          type: number
        duration:
          type: string
        thumbnailColor:
          type: string
      required:
        - id
        - title
        - publishedAt
        - views
        - likes
        - comments
        - watchTimeHours
        - engagementRate
        - duration
        - thumbnailColor

    OverviewStats:
      type: object
      description: |
        Aggregated KPI totals across tracked channels. totalSubscribers is the
        real sum of public subscriber counts. The 30D windowed totals
        (totalViews30d, totalWatchTimeHours30d, totalEstimatedRevenue30d) and
        the growth deltas (subscriberGrowth30d, viewsGrowth30d,
        revenueGrowth30d) are summed/computed only across channels that have
        Analytics API access via OAuth — they are null when no channel is
        connected. avgEngagementRate is computed from real recent video stats
        across all imported channels via the Data API.
      properties:
        totalSubscribers:
          type: integer
        totalViews30d:
          type: integer
          nullable: true
        totalWatchTimeHours30d:
          type: number
          nullable: true
        totalEstimatedRevenue30d:
          type: number
          nullable: true
        avgEngagementRate:
          type: number
          nullable: true
        subscriberGrowth30d:
          type: integer
          nullable: true
        viewsGrowth30d:
          type: number
          nullable: true
        revenueGrowth30d:
          type: number
          nullable: true
        topChannelByViews:
          type: string
          nullable: true
        topChannelByGrowth:
          type: string
          nullable: true
        oauthChannelCount:
          type: integer
          description: Number of tracked channels with per-brand OAuth (Analytics API access).
        totalChannelCount:
          type: integer
          description: Total number of tracked channels.
      required:
        - totalSubscribers
        - oauthChannelCount
        - totalChannelCount

    TrendPoint:
      type: object
      description: |
        A single day of aggregated metrics across OAuth-connected channels.
        Only real Analytics API data is ever included. If no channels have
        OAuth access, /overview/trends returns an empty array.
      properties:
        date:
          type: string
        totalViews:
          type: integer
        totalSubscribers:
          type: integer
        totalWatchTimeHours:
          type: number
        totalRevenue:
          type: number
      required:
        - date
        - totalViews
        - totalSubscribers
        - totalWatchTimeHours
        - totalRevenue